requirements.txt
transformers>=4.42
datasets>=2.20
accelerate>=0.33
peft>=0.11
evaluate>=0.4
lm-eval==0.4.2
bitsandbytes>=0.43 ; platform_system!="Darwin"
torch>=2.3
numpy>=1.26
pandas>=2.2
scipy>=1.11
matplotlib>=3.8
pyyaml>=6.0
tqdm>=4.66
wandb>=0.16
